third_party/blink/web_tests/fast/forms/ValidityState-typeMismatch-url-expected.txt

Input type=url validation test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS http://www.google.com is a correct valid url.
PASS http://foo:bar@www.google.com:80 is a correct valid url.
PASS http://localhost is a correct valid url.
PASS http://127.0.0.1 is a correct valid url.
PASS http://[0000:0000:0000:0000:0000:0000:7f00:0001]/ is a correct valid url.
PASS http://[0000:0000:0000:0000:0000:0000:127.0.0.1]/ is a correct valid url.
PASS http://[::7f00:0001]/ is a correct valid url.
PASS http://[1::2:3]/ is a correct valid url.
PASS http://[0000:0::ffff:10.0.0.1]/ is a correct valid url.
PASS http://a is a correct valid url.
PASS http://www.google.com/search?rls=en&q=WebKit&ie=UTF-8&oe=UTF-8 is a correct valid url.
PASS ftp://ftp.myhost.com is a correct valid url.
PASS ssh://ssh.myhost.com is a correct valid url.
PASS mailto:tkent@chromium.org is a correct valid url.
PASS mailto:tkent@chromium.org?body=hello is a correct valid url.
PASS file:///Users/tkent/.zshrc is a correct valid url.
PASS file:///C:/human.sys is a correct valid url.
PASS tel:+1-800-12345;ext=9999 is a correct valid url.
PASS tel:03(1234)5678 is a correct valid url.
PASS somescheme://ssh.myhost.com is a correct valid url.
PASS http://a/\/''*<>/ is a correct valid url.
PASS http://a/dfs/kds@sds is a correct valid url.
PASS http://a.a:1/search?a&b is a correct valid url.
PASS http://www.google.com/#top is a correct valid url.
PASS http://テスト。jp/パス?名前=値 is a correct valid url.
PASS www.google.com is a correct invalid url.
PASS 127.0.0.1 is a correct invalid url.
PASS .com is a correct invalid url.
PASS http://www.google.com:aaaa is a correct invalid url.
PASS :// is a correct invalid url.
PASS /http://www.google.com is a correct invalid url.
PASS ----ftp://a is a correct invalid url.
PASS scheme//a is a correct invalid url.
PASS http://[v8.:::]/ is a correct invalid url.
PASS http:// www.google.com is a correct valid url.
PASS http://www .google.com is a correct valid url.
PASS http://www.&#10;google.&#13;com is a correct valid url.
PASS http://host+ is a correct valid url.
PASS http://myurl! is a correct valid url.
PASS invalid is a correct invalid url.
PASS successfullyParsed is true

TEST COMPLETE
